+++ core/src/main/java/com/cognifide/aem/stubs/core/StubReload.java
@@ -1,9 +1,9 @@
 package com.cognifide.aem.stubs.core;
 
-import org.apache.commons.lang3.time.DurationFormatUtils;
-
 import java.text.NumberFormat;
 import java.util.Locale;
+
+import org.apache.commons.lang3.time.DurationFormatUtils;
 
 class StubReload {
 
@@ -34,9 +34,11 @@ class StubReload {
   }
 
   private String formatPercent(int value, int total) {
-    double percent = 0.0;
+    double percent;
     if (total > 0) {
       percent = ((double) value) / ((double) total);
+    } else {
+      percent = 0.0;
     }
     return NumberFormat.getPercentInstance(Locale.US).format(percent);
   }
